requirements.txt
streamlit==1.32.0
spacy==3.7.4
scikit-learn==1.4.1
PyMuPDF==1.24.0
python-docx==1.1.0
nltk==3.8.1
pandas==2.2.1
plotly==5.20.0
matplotlib==3.8.3
Pillow==10.2.0
